Continued work on transfer system refactor
- UploadBuffer object now has more responsibilities, including setting it's own internal data and managing the lifetime of it's contained allocation and buffer handle
- Use UploadBuffer to get data required for copy commands
- Remove unused functions from ResourceContextImpl
- Remove a bunch of stuff from TransferSystem, namely all our old functions, the custom pool usages, using it's own fence and command pool
- Move much of that responsibility to the TransferCommand class, which now manages lifetime of command buffer and attached pool + the related upload buffer + the reply
- TransferCommand will signal reply on completion using the fences, and will then clean itself up
- Implemented processing of messages in the TransferSystem, it now processes transfer messages and handles the submission of commands to the queue + waiting on them for a short duration
- Updated the command that forces transfers to complete to work properly, using fences to wait for each transfer to complete and making sure everything gets cleaned up
- Implemented basics of worker thread

+++ modules/resource_context/include/ResourceContext.hpp
@@ -15,7 +15,7 @@ struct ResourceContextCreateInfo
     bool validationEnabled;
 };
 
-constexpr static size_t k_TransferCompleteSemaphoreValue = 2u;
+constexpr static uint64_t k_TransferCompleteSemaphoreValue = 2u;
 
 class ResourceContextImpl;
